requirements.txt
streamlit
torch
torchvision
transformers
pillow
pandas
ultralytics
google-generativeai
scipy